duedates: extract routes into own classes
