--- slushy-app/src/main/java/net/kemitix/slushy/app/duedates/DueDateRoutes.java
@@ -1,37 +1,0 @@
-package net.kemitix.slushy.app.duedates;
-
-import org.apache.camel.builder.RouteBuilder;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-
-@ApplicationScoped
-public class DueDateRoutes
-        extends RouteBuilder {
-
-    @Inject SetDueInDays setDueInDays;
-    @Inject MarkCompleted markCompleted;
-
-    @Override
-    public void configure() {
-        from("direct:Slushy.DueIn30Days")
-                .routeId("Slushy.DueIn30Days")
-                .setHeader("SlushyDueInDays")
-                .constant("30")
-                .bean(setDueInDays)
-        ;
-
-        from("direct:Slushy.DueIn60Days")
-                .routeId("Slushy.DueIn60Days")
-                .setHeader("SlushyDueInDays")
-                .constant("60")
-                .bean(setDueInDays)
-        ;
-
-        from("direct:Slushy.DueCompleted")
-                .routeId("Slushy.DueCompleted")
-                .bean(markCompleted)
-        ;
-    }
-
-}
